refactor: extract decode graph warmup planning.
